feat(home): promote the story page to the default homepage
Move the /v2 route's data fetching and metadata to app/(home)/page.tsx
and delete /v2. The previous landing components remain in the repo for
a follow-up sweep.
