+++ app/(home)/page.tsx
@@ -1,14 +1,19 @@
-import Hero, { HeroBackground } from '@/components/landing/hero';
-import Paths from '@/components/landing/paths';
-import QuickLinks from '@/components/landing/quicklinks';
+import type { Metadata } from 'next';
+import StoryHome from '@/components/landing-v2/StoryHome';
 
 const BASE_URL = process.env.NEXT_PUBLIC_SITE_URL || 'https://build.avax.network';
+
+export const metadata: Metadata = {
+  title: 'Avalanche Builder Hub',
+  description:
+    'Sovereign L1 networks with sub-second finality, native interoperability, and protocol-level privacy and compliance controls.',
+};
 
 async function getGlobeData() {
   try {
     const [metricsRes, icmRes] = await Promise.all([
       fetch(`${BASE_URL}/api/overview-stats?timeRange=day`, {
-        next: { revalidate: 3600 }, // Cache for 1 hour
+        next: { revalidate: 3600 },
       }),
       fetch(`${BASE_URL}/api/icm-flow?days=30`, {
         next: { revalidate: 3600 },
@@ -30,17 +35,127 @@ async function getGlobeData() {
   }
 }
 
-export default async function HomePage(): Promise<React.ReactElement> {
-  const globeData = await getGlobeData();
+const PRIMARY_NETWORK_ID = '11111111111111111111111111111111LpoYY';
 
+// P-Chain-derived figures: L1 count (subnets flagged isL1 in the registry —
+// the Glacier activity aggregate undercounts at ~27) and the AVAX staked on
+// the primary network (economic security, the number institutions evaluate).
+async function getPChainStats(): Promise<{
+  l1Count: number | null;
+  primaryStakeAvax: number | null;
+}> {
+  try {
+    const res = await fetch(`${BASE_URL}/api/validator-stats?network=mainnet`, {
+      next: { revalidate: 3600 },
+    });
+    if (!res.ok) return { l1Count: null, primaryStakeAvax: null };
+    const subnets: { id: string; isL1?: boolean; totalStakeString?: string }[] = await res.json();
+    const count = subnets.filter((s) => s.isL1).length;
+    const primary = subnets.find((s) => s.id === PRIMARY_NETWORK_ID);
+    const stake = primary?.totalStakeString
+      ? Math.round(Number(BigInt(primary.totalStakeString) / 1_000_000_000n))
+      : null;
+    return { l1Count: count > 0 ? count : null, primaryStakeAvax: stake };
+  } catch (error) {
+    console.error('Failed to fetch P-Chain stats:', error);
+    return { l1Count: null, primaryStakeAvax: null };
+  }
+}
+
+// Live AVAX/USD for the stake figure — short revalidate keeps it honest,
+// and a failed fetch degrades to showing AVAX only (never a stale price).
+async function getAvaxUsdPrice(): Promise<number | null> {
+  try {
+    const res = await fetch(
+      'https://api.coingecko.com/api/v3/simple/price?ids=avalanche-2&vs_currencies=usd',
+      { headers: { accept: 'application/json' }, next: { revalidate: 300 } },
+    );
+    if (!res.ok) return null;
+    const data = await res.json();
+    const price = data?.['avalanche-2']?.usd;
+    return typeof price === 'number' && price > 0 ? price : null;
+  } catch (error) {
+    console.error('Failed to fetch AVAX price:', error);
+    return null;
+  }
+}
+
+// Circulating supply for the %-staked figure. Slow-moving, long revalidate.
+async function getCirculatingSupply(): Promise<number | null> {
+  try {
+    const res = await fetch(`${BASE_URL}/api/avax-supply`, { next: { revalidate: 14400 } });
+    if (!res.ok) return null;
+    const data = await res.json();
+    const supply = Number(data?.circulatingSupply);
+    return Number.isFinite(supply) && supply > 0 ? supply : null;
+  } catch (error) {
+    console.error('Failed to fetch AVAX supply:', error);
+    return null;
+  }
+}
+
+// DeFiLlama: capital metrics institutions screen for — TVL (committed
+// capital), stablecoin float (settlement liquidity), DEX volume (depth).
+async function getDefiStats(): Promise<{
+  tvlUsd: number | null;
+  stablesUsd: number | null;
+  dexVolume24hUsd: number | null;
+}> {
+  const opts = { next: { revalidate: 3600 } };
+  const [tvl, stables, dex] = await Promise.all([
+    fetch('https://api.llama.fi/v2/chains', opts)
+      .then((r) => (r.ok ? r.json() : null))
+      .then((rows) => rows?.find((c: any) => c.name === 'Avalanche')?.tvl ?? null)
+      .catch(() => null),
+    fetch('https://stablecoins.llama.fi/stablecoinchains', opts)
+      .then((r) => (r.ok ? r.json() : null))
+      .then((rows) => {
+        // sum every peg (USD, EUR, JPY, SGD, ...) — values are USD-denominated
+        const pegs = rows?.find((c: any) => c.name === 'Avalanche')?.totalCirculatingUSD;
+        if (!pegs) return null;
+        return Object.values(pegs).reduce(
+          (sum: number, v) => sum + (typeof v === 'number' && Number.isFinite(v) ? v : 0),
+          0,
+        );
+      })
+      .catch(() => null),
+    fetch(
+      'https://api.llama.fi/overview/dexs/avalanche?excludeTotalDataChart=true&excludeTotalDataChartBreakdown=true',
+      opts,
+    )
+      .then((r) => (r.ok ? r.json() : null))
+      .then((d) => d?.total24h ?? null)
+      .catch(() => null),
+  ]);
+  const num = (v: unknown) => (typeof v === 'number' && Number.isFinite(v) && v > 0 ? Math.round(v) : null);
+  return { tvlUsd: num(tvl), stablesUsd: num(stables), dexVolume24hUsd: num(dex) };
+}
+
+export default async function HomePage(): Promise<React.ReactElement> {
+  const [globeData, pchain, avaxUsd, circulatingSupply, defi] = await Promise.all([
+    getGlobeData(),
+    getPChainStats(),
+    getAvaxUsdPrice(),
+    getCirculatingSupply(),
+    getDefiStats(),
+  ]);
   return (
-    <>
-      <HeroBackground />
-      <Hero globeData={globeData} />
-      <main className="relative py-4 lg:py-16">
-        <QuickLinks />
-        {/* <Paths /> */}
-      </main>
-    </>
+    <StoryHome
+      globeData={globeData}
+      l1Count={pchain.l1Count}
+      primaryStakeAvax={pchain.primaryStakeAvax}
+      primaryStakeUsd={
+        pchain.primaryStakeAvax !== null && avaxUsd !== null
+          ? Math.round(pchain.primaryStakeAvax * avaxUsd)
+          : null
+      }
+      avaxUsdPrice={avaxUsd}
+      defi={defi}
+      supplyStakedPct={
+        pchain.primaryStakeAvax !== null && circulatingSupply !== null
+          ? (pchain.primaryStakeAvax / circulatingSupply) * 100
+          : null
+      }
+    />
   );
 }
